% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summaries.R
\name{aggregate_summaries}
\alias{aggregate_summaries}
\title{Utility function to aggregate summaries}
\usage{
aggregate_summaries(names, values)
}
\arguments{
\item{names}{Names of files that the summaries came from}

\item{values}{Summaries to collect.}
}
\description{
Utility function to aggregate summaries
}

