% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{clean_example_workflowHelper}
\alias{clean_example_workflowHelper}
\title{Function \code{clean_example_workflowHelper}}
\usage{
clean_example_workflowHelper(intern = F)
}
\arguments{
\item{intern}{Argument to system, option to capture output.}
}
\description{
Clean up files generated by the package example.
}

