% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{commands}
\alias{commands}
\title{Function \code{commands}}
\usage{
commands(...)
}
\arguments{
\item{...}{list of}
}
\value{
the \code{datasets}, \code{analyses}, \code{summaries}, \code{aggregates},
or \code{output} argument to \code{plan_workflow}
}
\description{
Turn a collection of R expressions into 
a character vector readable by \code{plan_workflow}. These
expressions stand for R commands to generate datasets, analyze
datasets, etc. You must specify the names of the fields.
For example, write \code{commands(x = data(y), z = 3)} instead of 
\code{commands(x, z)} or \code{commands(data(y), 3)}
}
\seealso{
\code{plan_workflow}
}

