% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{macro}
\alias{macro}
\title{Function \code{macro}}
\usage{
macro(arg)
}
\arguments{
\item{arg}{Character, name of the macro you want.}
}
\value{
The chosen macro.
}
\description{
Get a placeholder macro. These macros help
apply each analysis to each dataset and each summary to each
analysis of each dataset. Macros are not case-sensitive.
}

