% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dir.R
\name{new_dir}
\alias{new_dir}
\title{Function \code{new_dir}}
\usage{
new_dir(path)
}
\arguments{
\item{path}{path to directory}
}
\value{
directory name
}
\description{
make a new directory for storing simulations and results
}

