% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_aggregates}
\alias{parse_aggregates}
\title{Internal function}
\usage{
parse_aggregates(aggregates, summaries)
}
\arguments{
\item{aggregates}{Data frame about output files.}

\item{summaries}{Data frame of information about summaries to aggregate.}
}
\description{
Internal function
}

