% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_analyses}
\alias{parse_analyses}
\title{Internal function}
\usage{
parse_analyses(datasets = NULL, analyses = NULL)
}
\arguments{
\item{datasets}{Data frame of information targets to generate.}

\item{analyses}{Data frame of information targets to generate.}
}
\value{
A sanatized data frame
}
\description{
Internal function
}

