% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_output}
\alias{parse_output}
\title{Internal function}
\usage{
parse_output(output)
}
\arguments{
\item{output}{Data frame about output files.}
}
\value{
parsed data frame
}
\description{
Internal function
}

