% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_summaries.R
\name{parse_summaries}
\alias{parse_summaries}
\title{Internal function}
\usage{
parse_summaries(datasets = NULL, analyses = NULL, summaries = NULL)
}
\arguments{
\item{datasets}{Data frame of information targets to generate.}

\item{analyses}{Data frame of information targets to generate.}

\item{summaries}{Data frame of information targets to generate.}
}
\value{
A sanatized data frame.
}
\description{
Internal function
}

