% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholders.R
\name{placeholders}
\alias{placeholders}
\title{Function \code{placeholders}}
\usage{
placeholders()
}
\description{
Placeholder placeholders for commands.
}
\details{
These are placeholders to use when you're writing 
commands to generate data, run analyses, etc. They stand in 
for files, and they are for 
situations where the program can infer the correct file to
load, the correct file to save, etc. 
The save and load macros are for internal use only.
}

