% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_aggregate.R
\name{plan_aggregate}
\alias{plan_aggregate}
\title{Function \code{plan_aggregate}}
\usage{
plan_aggregate(sources, packages, type, summaries)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{type}{Character string, type of summary to aggregate}

\item{summaries}{Names of summaries}
}
\description{
Internal function to aggregate summaries
}

