% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_aggregates.R
\name{plan_aggregates}
\alias{plan_aggregates}
\title{Internal function}
\usage{
plan_aggregates(summaries, aggregates, sources, packages)
}
\arguments{
\item{summaries}{Data frame of information about summaries to aggregate.}

\item{aggregates}{Data frame about output files.}

\item{sources}{Character vector of R code files to load.}

\item{packages}{Character vector of R packages to load.}
}
\description{
Internal function
}

