% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_analysis.R
\name{plan_analysis}
\alias{plan_analysis}
\title{Function \code{plan_analysis}}
\usage{
plan_analysis(sources, packages, command, dataset, analysis)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character string of command to run.}

\item{dataset}{Character string, dataset to analyze}

\item{analysis}{Character string, analysis method}
}
\value{
name of analysis
}
\description{
Internal function to plan analysis
}

