% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_dataset.R
\name{plan_dataset}
\alias{plan_dataset}
\title{Function \code{plan_dataset}}
\usage{
plan_dataset(sources, packages, command, dataset)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character string, command to run.}

\item{dataset}{Character string, dataset to generate}
}
\description{
Internal function to plan dataset generation
}

