% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_output.R
\name{plan_output}
\alias{plan_output}
\title{Internal function}
\usage{
plan_output(output, sources, packages, depends)
}
\arguments{
\item{output}{Data frame about output to generate.}

\item{sources}{Character vector of R code files to load.}

\item{packages}{Character vector of R packages to load.}

\item{depends}{all possible RDS cached dependencies for output}
}
\description{
Internal function
}

