% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_stage.R
\name{plan_stage}
\alias{plan_stage}
\title{Internal function}
\usage{
plan_stage(df, sources, packages)
}
\arguments{
\item{df}{Data frame for producing targets.}

\item{sources}{Character vector of R code files to load.}

\item{packages}{Character vector of R packages to load.}
}
\description{
Internal function
}

