% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_summary.R
\name{plan_summary}
\alias{plan_summary}
\title{Function \code{plan_summary}}
\usage{
plan_summary(sources, packages, command, dataset, analysis, summary)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character string, command to run.}

\item{dataset}{Character string, dataset to analyze}

\item{analysis}{Character string, analysis method}

\item{summary}{Character string, summary method}
}
\description{
Internal function to summarize an analysis
}

