% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_workflow.R
\name{plan_workflow}
\alias{plan_workflow}
\title{Function \code{plan_workflow}}
\usage{
plan_workflow(sources, datasets = NULL, analyses = NULL, summaries = NULL,
  output = NULL, makefile = "Makefile", remakefile = "remake.yml",
  begin = NULL, clean = NULL, remake_args = list())
}
\arguments{
\item{sources}{Named character vector, code files and packages to load.
Code files should end in \code{.r} or \code{.R}. Otherwise, they will
be assumed to be packages.}

\item{datasets}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{analyses}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{summaries}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{output}{Named character vector of commands to make targets.
Names stand for files (possible non-RDS files) WITH their extensions.}

\item{makefile}{Character, name of the Makefile. Should be in the current
working directory.}

\item{remakefile}{Character, name of the \code{remake} file to generate. Should be in the current working directory.}

\item{begin}{Character vector of lines to prepend to the Makefile.}

\item{clean}{Character vector of extra shell commands for \code{make clean}.}

\item{remake_args}{Fully-named list of additional arguments to \code{remake::make}.
You cannot set \code{target_names} or \code{remake_file} this way.}
}
\description{
Main function of the package. Produces a Makefile to run a workflow.
}

