% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_assignment_from_command}
\alias{remove_assignment_from_command}
\title{Function \code{remove_assignment_from_command}}
\usage{
remove_assignment_from_command(cmd, x)
}
\arguments{
\item{cmd}{Command to clean}

\item{x}{Name of value assigned to by the command.}
}
\value{
parsed command
}
\description{
Special function to handle command syntax.
Turns "..plot.. <- mse_plot(mse)" into "mse_plot(mse)", etc.
}
\seealso{
\code{plan_workflow}
}

