% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{reps}
\alias{reps}
\title{Function \code{reps}}
\usage{
reps(x, reps = 2)
}
\arguments{
\item{x}{Character vector of commands, possibly returned from the
\code{commands} function or \code{strings} function.}

\item{reps}{Number of replicates.}
}
\value{
the \code{datasets}, \code{analyses}, \code{summaries}, \code{aggregates},
or \code{output} argument to \code{plan_workflow}
}
\description{
Replicate commands. For example, if
\code{x <- commands(data1 = rnorm(5), data2 = rpois(5,1))}, then
\code{reps(x, 2)} is
}
\seealso{
\code{commands}
}

