% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_example_workflow}
\alias{run_example_workflow}
\title{Function \code{run_example_workflow}}
\usage{
run_example_workflow()
}
\description{
Set up and run the package example: i.e., generate
the \code{remake} file and Makefile and run \code{make}.
}

