% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_example_workflowHelper}
\alias{run_example_workflowHelper}
\title{Function \code{run_example_workflowHelper}}
\usage{
run_example_workflowHelper(intern = F)
}
\arguments{
\item{intern}{TRUE/FALSE Argument to system to output to console or not.}
}
\description{
Set up and run the package example: i.e., generate
the \code{remake} file and Makefile and run \code{make}.
}

