% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{setup_example_workflowHelper}
\alias{setup_example_workflowHelper}
\title{Function \code{setup_example_workflow}}
\usage{
setup_example_workflowHelper()
}
\description{
Set up and run the package example: i.e., generate
the \code{remake} file and Makefile and run \code{make}.
}

