% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_step.R
\name{single_step}
\alias{single_step}
\title{Function \code{single_step}}
\usage{
single_step(sources, packages, command, root, dataset = "", analysis = "",
  summaries = "", append_rds = TRUE)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{dependsd character vector of commands to run.}

\item{root}{root of file}

\item{dataset}{name of RDS dataset file}

\item{analysis}{name of RDS analysis file}

\item{summaries}{Character vector, names of RDS summary files}

\item{append_rds}{logical, whether to append a \code{.rds} file extension to the output}
}
\description{
Plans a stage of the workflow. Utility function of \code{plan_workflow}.
}
\seealso{
\code{plan_workflow}
}

