% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{write_example_workflowHelper}
\alias{write_example_workflowHelper}
\title{Function \code{write_example_workflow}}
\usage{
write_example_workflowHelper()
}
\description{
Write files to generate the example workflow: i.e., 
produce \code{code.R} and \code{Makefile.R}
}

