% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_summaries.R
\name{aggregate_summaries}
\alias{aggregate_summaries}
\title{Internal function}
\usage{
aggregate_summaries(names, values)
}
\arguments{
\item{names}{Names of the output list.}

\item{values}{Values of the output list}
}
\value{
A named list
}
\description{
Internal function
}

