% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knitr_depends}
\alias{knitr_depends}
\title{Function \code{knitr_depends}}
\usage{
knitr_depends(output, all_depends)
}
\arguments{
\item{output}{Data frame of "output" stage targets}

\item{all_depends}{Dependencies of fake target `all`.}
}
\value{
knitr dependencies
}
\description{
Resolve dependencies of knitr targets: 
i.e., everything except the knitr targets themeslves 
and everything downstream.
}

