% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_aggregate.R
\name{plan_aggregate}
\alias{plan_aggregate}
\title{Function \code{plan_aggregate}}
\usage{
plan_aggregate(sources, packages, command, aggregate, summaries)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character string, command to run}

\item{aggregate}{Character string, aggretation method}

\item{summaries}{Character vector, methods of summarizing analysis}
}
\description{
Internal function to aggregate summaries
}

