% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_output.R
\name{plan_output}
\alias{plan_output}
\title{Function \code{plan_output}}
\usage{
plan_output(sources, packages, command, output)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character vector of commands to run.}

\item{output}{Character string, output to compute}
}
\value{
name of output
}
\description{
Internal function to compute final output
}

