% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_workflow.R
\name{plan_workflow}
\alias{plan_workflow}
\title{Function \code{plan_workflow}}
\usage{
plan_workflow(sources, datasets = NULL, analyses = NULL, summaries = NULL,
  output = NULL, clean = NULL, makefile = "Makefile")
}
\arguments{
\item{sources}{Named character vector, code files and packages to load.
Code files should end in \code{.r} or \code{.R}. Otherwise, they will
be assumed to be packages.}

\item{datasets}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{analyses}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{summaries}{Named character vector of commands to make targets.
Names stand for RDS target files without their extensions.}

\item{output}{Named character vector of commands to make targets.
Names stand for files (possible non-RDS files) WITH their extensions.}

\item{clean}{Character vector of extra shell commands for \code{make clean}.}

\item{makefile}{Character, name of the Makefile. Should be in the current
working directory. Otherwise, the \code{YAML} files will not be found
and the \code{make} will not work.}
}
\description{
Main function of the package. Produces a Makefile to run an workflow.
}

