% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_workflow.R
\name{plan_workflow}
\alias{plan_workflow}
\title{Function \code{plan_workflow}}
\usage{
plan_workflow(sources, datasets = NULL, analyses = NULL, summaries = NULL,
  output = NULL)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code 
(*.R and *.r files) along with external packages (no file extension).}

\item{datasets}{Named character vector of commands to generate datasets.}

\item{analyses}{Named character vector of commands to analyze the datasets.}

\item{summaries}{Named character vector of commands to condense the data and 
analyses into small summaries.}

\item{output}{Named character vector, commands for general output. Unlike the 
other command vectors, the names here need not denote rds files. However, they 
should have the appropriate file extensions.}
}
\description{
Plan a simulation study workflow or something similar. Produces a makefile 
so that you can run or resume you workflow with `make` or `make -j <whatever>` for
parallelization.
}

