% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_step.R
\name{single_step}
\alias{single_step}
\title{Function \code{single_step}}
\usage{
single_step(sources, packages, command, file, yaml)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{dependsd character vector of commands to run.}

\item{file}{target file to make}

\item{yaml}{YAML/remake file to use}
}
\description{
Internal function to plan a stage of the workflow. 
Utility function of \code{plan_workflow}.
}
\seealso{
\code{plan_workflow}
}

