% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_step.R
\name{single_step}
\alias{single_step}
\title{Function \code{single_step}}
\usage{
single_step(sources, packages, command, save, depends = NULL,
  replacements = NULL)
}
\arguments{
\item{sources}{Character vector of paths to the files containing your R code.}

\item{packages}{Character vector of packages that your code depends on.}

\item{command}{Character, command to run.}

\item{save}{Character, target file to save}

\item{depends}{named character vector of file dependencies}

\item{replacements}{Named character vector, substitutions to make in command placeholders}
}
\value{
list of fields for the YAML file
}
\description{
Internal function to plan a stage of the workflow. 
Utility function of \code{plan_workflow}.
}
\seealso{
\code{plan_workflow}
}

